+++ core/resources/views/master/inc/normal.blade.php
@@ -224,8 +224,7 @@
                         <li><a class="sub-link" href="{{ route('back.staff.index') }}"><span class="sub-item">{{ __('System User') }}</span></a></li>
                     @endif
                     @if ($can('System Backup'))
-                        <li><a class="sub-link" href="{{ route('back.system.backup') }}"><span class="sub-item">{{ __('System Backup') }}</span></a></li>
-                        <li><a class="sub-link" href="{{ route('back.database.backup') }}"><span class="sub-item">{{ __('Database Backup') }}</span></a></li>
+                        <li><a class="sub-link" href="{{ route('back.backup.index') }}"><span class="sub-item">Backups do sistema</span></a></li>
                     @endif
                     <li><a class="sub-link" href="{{ route('front.cache.clear') }}"><span class="sub-item">{{ __('Cache Clear') }}</span></a></li>
                 </ul>